Waterwheel provider for Vue.
@TODO: Move Waterwheel provider to it's own project so it can be
included as this should be a dependency not something managed in
project.

Updates the waterwheel provider so that it works with options on the
vue component.

```
import { WaterwheelProvider } from './services/waterwheel.service'

new Vue({
  provider: WaterwheelProvider.provide()
})
```

Component can specify the `waterwheel` option with an object keyed to
the return value of the requests.

``` js

waterwheel: {
  dogs: 'node/dogs'
}

```

Will trigger an API lookup when the component is created in Vue and will
make the resulting API call available via `{{ dogs }}`, the value should
be defaulted in your `data()` function for the component.

Only fetching is currently supported.